Fixed all compiler warnings

+++ Hoowan/src/Hoowan/Renderer/Renderer2D.cpp
@@ -132,7 +132,7 @@ namespace Hoowan
 		HW_PROFILE_FUNCTION();
 
 		// Size needs to be casted down to something that is one byte to get element numbers
-		uint32_t dataSize = (uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase;
+		uint32_t dataSize = (uint32_t)((uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase);
 		s_Data.QuadVertexBuffer->SetData(s_Data.QuadVertexBufferBase, dataSize);
 
 		Flush();
